requirement.txt
langchain
python-dotenv
langchain-openai
ipykernel
langchain_openai
langchain-core



